requirements.txt
streamlit
pandas
matplotlib
seaborn

langchain
langchain-community
langchain-huggingface
python-dotenv
faiss-cpu
